requirements.txt
pip==24.3.1
addict==2.4.0
altair==5.4.1
attrs==24.2.0
bleach==6.2.0
blinker==1.8.2
bokeh==3.6.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
configargparse==1.7
contourpy==1.3.0
cycler==0.12.1
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
fastjsonschema==2.20.0
filelock==3.16.1
flask==3.0.3
fonttools==4.54.1
fsspec==2024.10.0
gitdb==4.0.11
gitpython==3.1.43
idna==3.10
imageio==2.36.0
importlib-resources==6.4.5
ipywidgets==8.1.5
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab-widgets==3.0.13
kiwisolver==1.4.7
lazy-loader==0.4
linkify-it-py==2.0.3
lxml==5.3.0
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.9.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.13.3
nbformat==5.10.4
networkx==3.4.2
nibabel==5.3.2
nilearn==0.10.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
pandas==2.2.3
panel==1.5.3
param==2.1.1
patsy==0.5.6
pillow==11.0.0
plotly==5.24.1
pooch==1.8.2
protobuf==5.28.3
pyarrow==18.0.0
pydeck==0.9.1
pyparsing==3.2.0
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
pyvista==0.44.1
pyviz-comms==3.0.3
pyyaml==6.0.2
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rpds-py==0.20.1
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
scooby==0.10.0
seaborn==0.13.2
simpleitk==2.4.0
smmap==5.0.1
stack-data==0.6.3
statsmodels==0.14.4
stpyvista==0.1.4
streamlit==1.40.0
streamlit-option-menu==0.4.0
SimpleITK ==2.4.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2024.9.20
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.66.6
trimesh==4.5.1
triton==3.1.0
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
vtk==9.3.1
watchdog==5.0.3
webencodings==0.5.1
werkzeug==3.0.6
widgetsnbextension==4.0.13
xyzservices==2024.9.0
